requirements.txt
Flask==3.1.2
matplotlib==3.10.7
numpy==2.3.3
openai==2.3.0
pandas==2.3.3
python-dotenv==1.1.1
seaborn==0.13.2
selenium==4.36.0
Werkzeug==3.1.3
